Updated formatting + added a "Read_Point" method
Almost all of the changes in this commit are formatting ones. Atom automatically formats new lines (now that it knows that I'm dealing with a python file)... this format is different from the one that the skeleton used... so, I updated the skeleton formatting

I also added an is_equal method (for floating point equality), filled in the rectangle/circle methods that hadn't been defined, and added a function to read in a point from the file. See Geom.py:209.

In the next commit, I should have the rest of the printouts going (in other words, I'll have a finished draft of the code).