Jenny suggestions.
I expanded a few comments and rearranged some stuff. I replaced some dumb code in circle_circumscribe (line 118). In partiuclar, center_x and center_y are the average of the edges of the rectangle (see line 120 for explanation).

I also moved the interval_intersection method into the Rectangle class as a private method.

The code gives the same results right now as it use to, so I think it's good to go.