Modified how circumscribe methods work + added tests
After talking with the TA's, I decided to change circle_circumscribe and rectangle_circumscribe so that they modify self rather than returning the desired rectangle/circle.

This makes the code a little easier to understand (I think).

I also added tests for the Point and Circle classes. I plan on added tests for the Rectangle class in the next commit. So far, everything is working (except for the occasional glitch associated with floating point nonsense).